feat(ios): raster-layer effects, Gaussian Blur/Feather, and the Beam Dither port (v2.33.0) (#388)
* feat(ios): draw Gaussian Blur and Feather (#372 part 2)

The desktop's two *fill-replacing* effects: instead of adding a coloured plane
beside the layer, they blur its premultiplied RGBA and hand back a replacement
fill. iOS kept them in the document but drew the fill unblurred, and the sheet
said so.

`AppearanceRenderer.blurRGBA` is `npimage.gaussian_blur` — a triple box blur
over all four channels — and it is BANDED, which is the answer to the memory
question #372 asked. The issue offered three options: bands, a reduced scale,
or refusing. The desktop had already hit this wall in #347 and taken the first,
and its reasoning ports exactly: the three vertical passes are exact Int64
window sums, so where a band boundary falls cannot change the result, and the
horizontal passes are per-row and cannot see a boundary at all. The transient
is bounded by the radius rather than by the picture, and the output is
identical — so desktop parity survives, which a reduced-scale render would
have cost.

`featherRGBA` keeps the lesser of the blurred and original alpha. That minimum
is the whole difference from a blur: a blur grows the silhouette, feather may
only eat into it, so a feathered layer stays inside its own bounds. Colour is
rescaled to the surviving alpha because the buffer is premultiplied, in Double
with a truncating cast — numpy promotes uint32/float32 to float64 and
astype(uint32) truncates, and a rounded Float here disagrees by one LSB.

The stack is now walked IN ORDER, because order is meaningful: a blur below a
drop shadow means the shadow is cast by the blurred silhouette; the same blur
above it softens only the artwork and leaves the shadow sharp. Resolving every
plane against one alpha, as before, silently picks one reading and is wrong
about the other. Padding is applied once up front rather than per effect —
equivalent, because the box sums truncate at the array edge with zero outside,
which *is* zero-padding — which keeps every plane and the replacement fill in
one coordinate space.

Both compositors draw it: the canvas, and ProjectArchive's separate preview
loop, which has its own test because it scales and would otherwise preview a
blurred layer as a sharp one.

Testing: 309 iOS unit tests green, 9 new. scripts/gen-fill-override-fixture.py
runs the desktop's own npimage.gaussian_blur and the feather branch over a
fixture whose radii make the halo span several iOS bands, and the Swift side
matches it word for word — a banding mistake cannot hide behind a small
picture. Separate tests pin that feather never grows the silhouette, that a
row-uniform image blurs to a row-uniform one (a band seam is exactly a row
that disagrees with its neighbours), and that stack order changes the result.

Still open on #372: the Effects sheet on raster and photo layers.


* feat(ios): raster-layer effects and the Beam Dither port (v2.33.0)

Closes #372 and #385, which clears every backlog item that is mine to close.

## Effects on raster and photo layers (#372 part 1)

The model, the archive and the renderer were always layer-agnostic; only the
entry point was text-only. What actually had to be decided was memory: a text
layer's effect planes are a tight box around its glyphs, while a photo layer's
are the whole canvas, which is the jetsam risk #309 and #311 were about.

The stack is measured — `AppearanceRenderer.transientLayers` counts the alpha
plane, the working copy, one colour buffer per plane, and the padded source a
fill override needs — and put through the same budget every other allocation
door in the app uses since #354. Refused out loud beats discovered as a jetsam
kill with the document unsaved.

## Beam Dither on iOS (#385)

`Dither.swift` ports `photoslop.dither`: six error-diffusion kernels scanned
serpentine, the Bayer doubling recurrence, and the beam model. It is proven
against the desktop TWICE — `scripts/gen-dither-fixture.py` for the engine
(every kernel at 2 and 4 levels, three Bayer sizes, three beam configurations)
and seven new `gen-filter-fixture.py` cases for the whole filter, word for
word, which is the same bar the other seven filters meet.

Everything is `Double`, because a dither turns a rounding difference into a
different pixel being lit. `rint` breaks ties to even, as Python's `round` and
numpy's `rint` do, and there is a test for it: 0.5 is exactly what a mid-grey
plane hands a two-level quantiser.

Error diffusion is sequential by nature — a pixel cannot be quantised until its
left neighbour has pushed error into it — so unlike the other seven filters it
does not vectorise over `PixelBuffer`. Cell size is the cost control, and the
filter declares its working buffers to the memory budget.

Free-text `#RRGGBB` filter parameters are new, for the tri-tone inks; the sheet
renders them as text fields and a malformed colour falls back rather than
aborting a render in progress.

## Two precision cliffs in the desktop, found by the port

Neither was visible until two implementations had to agree.

- `dither.beam_mask` built its row index as `float32`, throwing away ~1e-7 of a
  phase that a cosine turns into a coverage value.
- `dither.error_diffuse` returned a `float32` plane while every other function
  in the module returns `float64`. At two levels its values are 0 and 1, which
  float32 holds exactly — which is why it looked fine — but at four levels a
  third came back as 0.3333333432674408.
- `filters.BeamDitherFilter` conditioned in `float32` for the same reason, and
  now does not: every stage in this pipeline feeds a quantiser.

Each is invisible in a single render and each is exactly enough to light a
different pixel, which is why the iOS port could not reproduce the engine until
they were removed.

Testing: 330 iOS unit tests green, 21 new. 3667 Python tests green. Ruff clean,
version gate 2.33.0.


---------

+++ photoslop/dither.py
@@ -145,7 +145,11 @@ def error_diffuse(lum: np.ndarray, kernel: str, levels: int = 2) -> np.ndarray:
     steps = max(2, int(levels)) - 1
 
     rows = [row.tolist() for row in lum.astype(np.float64)]
-    out = np.empty((height, width), dtype=np.float32)
+    # float64 like every other function here returns. float32 was the odd one
+    # out and quietly lossy: at levels=2 its only values are 0 and 1, which it
+    # holds exactly, but at 4 levels a third comes back as 0.3333333432674408
+    # and the plane no longer means quite what the arithmetic computed.
+    out = np.empty((height, width), dtype=np.float64)
     for y in range(height):
         row = rows[y]
         rightwards = y % 2 == 0
@@ -212,7 +216,12 @@ def beam_mask(lum: np.ndarray, pitch: int, amplitude: float, bend: float = 1.0) 
     """
     height, width = lum.shape
     pitch = max(2, int(pitch))
-    rows = np.arange(height, dtype=np.float32).reshape(-1, 1)
+    # float64, not float32: the row index feeds a phase that a cosine turns
+    # into a coverage value, and a float32 divide here throws away about 1e-7
+    # of it for nothing. That is invisible in a single render and is exactly
+    # enough to flip a stippled pixel, which is what stopped the iOS port
+    # (#385) from reproducing this engine until the cliff was removed.
+    rows = np.arange(height, dtype=np.float64).reshape(-1, 1)
     # Phase along the vertical raster, advanced by the signal. `bend` lets the
     # deflection be dialled out entirely (straight scanlines) without losing
     # the intensity modulation.
